+++ dozer-types/src/models/app_config.rs
@@ -93,7 +93,7 @@ pub struct Config {
     #[prost(message, tag = "18")]
     /// Dozer Cloud specific configuration
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub err_threshold: Option<u64>,
+    pub err_threshold: Option<u32>,
 }
 
 pub fn default_home_dir() -> String {
@@ -131,7 +131,7 @@ pub fn default_commit_timeout() -> u64 {
     50
 }
 
-pub fn default_err_threshold() -> u64 {
+pub fn default_err_threshold() -> u32 {
     0
 }
 
@@ -196,7 +196,7 @@ impl<'de> Deserialize<'de> for Config {
                 let mut app_buffer_size: Option<u32> = Some(default_app_buffer_size());
                 let mut commit_size: Option<u32> = Some(default_commit_size());
                 let mut commit_timeout: Option<u64> = Some(default_commit_timeout());
-                let mut err_threshold: Option<u64> = Some(default_err_threshold());
+                let mut err_threshold: Option<u32> = Some(default_err_threshold());
 
                 while let Some(key) = access.next_key()? {
                     match key {
@@ -249,7 +249,7 @@ impl<'de> Deserialize<'de> for Config {
                             cloud = access.next_value::<Option<Cloud>>()?;
                         }
                         "err_threshold" => {
-                            err_threshold = access.next_value::<Option<u64>>()?;
+                            err_threshold = access.next_value::<Option<u32>>()?;
                         }
                         _ => {
                             access.next_value::<IgnoredAny>()?;
